Rename directory orm to database.
