+++ store/index.js
@@ -1,5 +1,5 @@
 import VuexORM from '@vuex-orm/core'
-import database from '../orm/database'
+import database from '../database/database'
 
 export const plugins = [ VuexORM.install(database) ]
